Update AS and fix SMPS2ASM sanity checks.
These sanity checks were using stale forward references, which can
cause bogus errors, as described here:
https://github.com/flamewing/asl-releases/issues/32

The new AS correctly detects and prevents this behaviour, and the
sanity checks have been rewritten to avoid invoking it in the first
place.

The builds of AS for platforms other than Windows still need to be
updated.